full content in rss description tag | sync hugo rss template changes
Full content in description tag
-------------------------------
Looks like there is no one correct way to add the full content to an
rss feed, which is one of the reasons why Hugo doesn't add it to the
default template (See gohugoio/hugo#4071). The two ways are either to
add a separate content:encoded tag or to just add it in the
description. Going with the separate description tag for now, as the
rss spec suggests to use it for full content for publishers who don't
want to employ summaries, and I'm such a publisher:
https://www.rssboard.org/rss-profile#namespace-elements-content-encoded

Another option here could be to switch to an atom feed instead, which
seems to have a more well-defined spec. Leaving this decision for
another day as Hugo has rss builtin for now, although adding an atom
feed is almost trivial as well. See: gohugoio/hugo#3473

Sync hugo rss template changes
------------------------------
Looks like the rss.xml file in this theme was just copied over from the
corresponding file in hugo repo at that time, with just content tag
added to it. Since then a lot of changes have been introduced in the
hugo version of the template, so synced those here.